a basic to do app, vanilla js, but with local storage setting
My favourite code yet. It challenged me a lot, still does. Trying to figure out how to getItem before setting it, then appending it to a list created as an element in js. Problem is, the code crashes when I try to follow whatever idea of doing what I just said. I am figuring it soon. Also has lots of commenting in the js. Its best for learning